[TACHYON-1069] Move assertValidPort to NetworkAddressUtils
