+++ servers/src/main/java/tachyon/worker/nio/NIODataServer.java
@@ -38,6 +38,7 @@ import tachyon.Constants;
 import tachyon.Sessions;
 import tachyon.conf.TachyonConf;
 import tachyon.exception.NotFoundException;
+import tachyon.util.network.NetworkAddressUtils;
 import tachyon.worker.DataServer;
 import tachyon.worker.DataServerMessage;
 import tachyon.worker.block.BlockDataManager;
@@ -85,7 +86,7 @@ public final class NIODataServer implements Runnable, DataServer {
       TachyonConf tachyonConf) {
     LOG.info("Starting DataServer @ " + address);
     mTachyonConf = Preconditions.checkNotNull(tachyonConf);
-    TachyonConf.assertValidPort(Preconditions.checkNotNull(address), mTachyonConf);
+    NetworkAddressUtils.assertValidPort(Preconditions.checkNotNull(address), mTachyonConf);
     mAddress = address;
     mDataManager = Preconditions.checkNotNull(dataManager);
     try {
